Revert "[Indigo] Velocity interface controller indigo"
