+++ cob_collision_velocity_filter/package.xml
@@ -1,6 +1,6 @@
 <package>
   <name>cob_collision_velocity_filter</name>
-  <version>0.5.4</version>
+  <version>0.5.3</version>
   <description>The cob_collision_velocity_filter package is a package for collision avoidance using teleoperation.</description>
  
   <license>LGPL</license>
@@ -10,7 +10,8 @@
 
   <maintainer email="mig@ipa.fhg.de">Matthias Gruhler</maintainer>
   <author>Matthias Gruhler</author>
-  <author>Michal Spanel</author>
+  <author>Michal Spanel (spanel@fit.vutbr.cz</author>
+  <author>vel. limited marker)</author>
 
   <buildtool_depend>catkin</buildtool_depend>
   <build_depend>roscpp</build_depend>
