--- cob_control/CHANGELOG.rst
@@ -1,8 +1,0 @@
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-Changelog for package cob_control
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-
-0.5.4 (2014-08-26)
-------------------
-* added meta-package
-* Contributors: ipa-fxm
