+++ cob_control/package.xml
@@ -1,6 +1,6 @@
 <package>
   <name>cob_control</name>
-  <version>0.5.4</version>
+  <version>0.5.0</version>
   <description>cob_control meta-package</description>
 
 
@@ -17,15 +17,10 @@
   <buildtool_depend>catkin</buildtool_depend>
   <run_depend>cob_base_velocity_smoother</run_depend>
   <run_depend>cob_collision_velocity_filter</run_depend>
-  <run_depend>cob_control_mode_adapter</run_depend>
   <run_depend>cob_footprint_observer</run_depend>
-  <run_depend>cob_frame_tracker</run_depend>
-  <run_depend>cob_hardware_interface</run_depend>
   <run_depend>cob_lookat_controller</run_depend>
-  <run_depend>cob_path_broadcaster</run_depend>
   <run_depend>cob_trajectory_controller</run_depend>
   <run_depend>cob_twist_controller</run_depend>
-
