--- cob_frame_tracker/package.xml
@@ -1,36 +1,0 @@
-<?xml version="1.0"?>
-<package>
-  <name>cob_frame_tracker</name>
-  <version>0.5.0</version>
-  <description>The cob_frame_tracker package</description>
-
-  <author email="fxm@ipa.fhg.de">Felix Messmer</author>
-  <maintainer email="fxm@ipa.fhg.de">Felix Messmer</maintainer>
-
-  <license>LGPL</license>
-
-  <buildtool_depend>catkin</buildtool_depend>
-  <build_depend>roscpp</build_depend>
-  <build_depend>geometry_msgs</build_depend>
-  <build_depend>std_srvs</build_depend>
-  <build_depend>cob_srvs</build_depend>
-  <build_depend>tf</build_depend>
-  <build_depend>tf_conversions</build_depend>
-  <build_depend>control_toolbox</build_depend>
-  
-  <run_depend>rospy</run_depend>
-  <run_depend>roscpp</run_depend>
-  <run_depend>geometry_msgs</run_depend>
-  <run_depend>std_srvs</run_depend>
-  <run_depend>cob_srvs</run_depend>
-  <run_depend>tf</run_depend>
-  <run_depend>tf_conversions</run_depend>
-  <run_depend>visualization_msgs</run_depend>
-  <run_depend>interactive_markers</run_depend>
-  <run_depend>control_toolbox</run_depend>
-
-
-  <export>
-
-  </export>
-</package>
