--- cob_hardware_interface/package.xml
@@ -1,33 +1,0 @@
-<?xml version="1.0"?>
-<package>
-  <name>cob_hardware_interface</name>
-  <version>0.5.0</version>
-  <description>
-    This package implements the cob_hardware_interface.
-    The interface uses topic communication rather than direct pointer access to the drivers.
-    I.e. it publishs the desired velocity to a topic and reads from the /joint_states topic.
-    Thus, it can be used for both velocity-controlled real hardware as well as velocity-controlled simulation (i.e. using the JointVelocityControllers)
-  </description>
-
-  <maintainer email="fxm@ipa.fhg.de">Felix Messmer</maintainer>
-  <author email="fxm@ipa.fhg.de">Felix Messmer</author>
-  
-  <license>BSD</license>
-
-
-  <buildtool_depend>catkin</buildtool_depend>
-  
-  <build_depend>hardware_interface</build_depend>
-  <build_depend>controller_manager</build_depend>
-  <build_depend>brics_actuator</build_depend>
-  
-  
-  <run_depend>hardware_interface</run_depend>
-  <run_depend>controller_manager</run_depend>
-  <run_depend>brics_actuator</run_depend>
-
-
-  <export>
-
-  </export>
-</package>
