--- cob_lookat_controller/CHANGELOG.rst
@@ -1,11 +1,0 @@
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-Changelog for package cob_lookat_controller
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-
-0.5.4 (2014-08-26)
-------------------
-* added explicit default argument queue_size
-* move launch files to cob_robots
-* fix dependency-hell on multiple cores
-* moved cob_lookat_controller
-* Contributors: Alexander Bubeck, Felix Messmer, ipa-fxm
