+++ cob_lookat_controller/CMakeLists.txt
@@ -1,29 +1,102 @@
 cmake_minimum_required(VERSION 2.8.3)
 project(cob_lookat_controller)
 
-# Load catkin and all dependencies required for this package
-find_package(catkin REQUIRED COMPONENTS roscpp sensor_msgs geometry_msgs brics_actuator kdl_parser kdl_conversions tf tf_conversions )
+## Find catkin macros and libraries
+## if COMPONENTS list like find_package(catkin REQUIRED COMPONENTS xyz)
+## is used, also find other catkin packages
+find_package(catkin REQUIRED COMPONENTS roscpp sensor_msgs geometry_msgs std_srvs cob_srvs brics_actuator kdl_parser kdl_conversions tf tf_conversions )
 
+## System dependencies are found with CMake's conventions
+# find_package(Boost REQUIRED COMPONENTS system)
 find_package(orocos_kdl REQUIRED)
 
-# Declare catkin package
+## Uncomment this if the package has a setup.py. This macro ensures
+## modules and global scripts declared therein get installed
+## See http://ros.org/doc/api/catkin/html/user_guide/setup_dot_py.html
+# catkin_python_setup()
+
+################################################
+## Declare ROS messages, services and actions ##
+################################################
+
+## To declare and build messages, services or actions from within this
+## package, follow these steps:
+## * Let MSG_DEP_SET be the set of packages whose message types you use in
+##   your messages/services/actions (e.g. std_msgs, actionlib_msgs, ...).
+## * In the file package.xml:
+##   * add a build_depend and a run_depend tag for each package in MSG_DEP_SET
+##   * If MSG_DEP_SET isn't empty the following dependencies might have been
+##     pulled in transitively but can be declared for certainty nonetheless:
+##     * add a build_depend tag for "message_generation"
+##     * add a run_depend tag for "message_runtime"
+## * In this file (CMakeLists.txt):
+##   * add "message_generation" and every package in MSG_DEP_SET to
+##     find_package(catkin REQUIRED COMPONENTS ...)
+##   * add "message_runtime" and every package in MSG_DEP_SET to
+##     catkin_package(CATKIN_DEPENDS ...)
+##   * uncomment the add_*_files sections below as needed
+##     and list every .msg/.srv/.action file to be processed
+##   * uncomment the generate_messages entry below
+##   * add every package in MSG_DEP_SET to generate_messages(DEPENDENCIES ...)
+
+## Generate messages in the 'msg' folder
+# add_message_files(
+#   FILES
+#   Message1.msg
+#   Message2.msg
+# )
+
+## Generate services in the 'srv' folder
+# add_service_files(
+#   FILES
+#   Service1.srv
+#   Service2.srv
+# )
+
+## Generate actions in the 'action' folder
+# add_action_files(
+#   FILES
+#   Action1.action
+#   Action2.action
+# )
+
+## Generate added messages and services with any dependencies listed here
+# generate_messages(
+#   DEPENDENCIES
+#   std_msgs  # Or other packages containing msgs
+# )
+
+###################################
+## catkin specific configuration ##
+###################################
+## The catkin_package macro generates cmake config files for your package
+## Declare things to be passed to dependent projects
+## INCLUDE_DIRS: uncomment this if you package contains header files
+## LIBRARIES: libraries you create in this project that dependent projects also need
+## CATKIN_DEPENDS: catkin_packages dependent projects also need
+## DEPENDS: system dependencies of this project that dependent projects also need
 catkin_package(
-  CATKIN_DEPENDS roscpp sensor_msgs geometry_msgs brics_actuator
   INCLUDE_DIRS include
-  LIBRARIES cob_lookat_driver cob_lookat_controller
+  LIBRARIES cob_lookat_driver cob_lookat_controller cob_frame_tracker
+  CATKIN_DEPENDS roscpp sensor_msgs geometry_msgs std_srvs cob_srvs brics_actuator
+#  DEPENDS system_lib
 )
 
-include_directories(include ${catkin_INCLUDE_DIRS} ${orocos_kdl_INCLUDE_DIRS})
+###########
+## Build ##
+###########
 
+## Specify additional locations of header files
+## Your package locations should be listed before other locations
+include_directories(include)
+include_directories(${catkin_INCLUDE_DIRS})
+include_directories(${orocos_kdl_INCLUDE_DIRS})
 
-## Build
-# Declare a cpp library
+## Declare a cpp library
 add_library(cob_lookat_driver src/cob_lookat_driver.cpp )
-add_dependencies(cob_lookat_driver brics_actuator_gencpp)
-
 add_library(cob_lookat_controller src/cob_lookat_controller.cpp )
-add_dependencies(cob_lookat_controller brics_actuator_gencpp)
 target_link_libraries(cob_lookat_controller ${catkin_LIBRARIES} ${orocos_kdl_LIBRARIES})
+add_library(cob_frame_tracker src/cob_frame_tracker.cpp )
 
 ## Declare a cpp executable
 add_executable(cob_lookat_driver_node src/cob_lookat_driver_node.cpp)
@@ -32,10 +105,25 @@ target_link_libraries(cob_lookat_driver_node cob_lookat_driver ${catkin_LIBRARIE
 add_executable(cob_lookat_controller_node src/cob_lookat_controller_node.cpp)
 target_link_libraries(cob_lookat_controller_node cob_lookat_controller ${catkin_LIBRARIES} ${orocos_kdl_LIBRARIES})
 
+add_executable(cob_frame_tracker_node src/cob_frame_tracker_node.cpp)
+target_link_libraries(cob_frame_tracker_node cob_frame_tracker ${catkin_LIBRARIES})
 
-## Install
-# Mark executables and/or libraries for installation
-install(TARGETS cob_lookat_driver cob_lookat_driver_node cob_lookat_controller cob_lookat_controller_node
+#############
+## Install ##
+#############
+
+# all install targets should use catkin DESTINATION variables
+# See http://ros.org/doc/api/catkin/html/adv_user_guide/variables.html
+
+## Mark executable scripts (Python etc.) for installation
+## in contrast to setup.py, you can choose the destination
+# install(PROGRAMS
+#   scripts/my_python_script
+#   DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION}
+# )
+
+## Mark executables and/or libraries for installation
+install(TARGETS cob_lookat_driver cob_lookat_driver_node cob_lookat_controller cob_lookat_controller_node cob_frame_tracker cob_frame_tracker_node
  ARCHIVE DESTINATION ${CATKIN_PACKAGE_LIB_DESTINATION}
  LIBRARY DESTINATION ${CATKIN_PACKAGE_LIB_DESTINATION}
  RUNTIME DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION}
@@ -48,7 +136,22 @@ install(DIRECTORY include/${PROJECT_NAME}/
  PATTERN ".svn" EXCLUDE
 )
 
-# Mark other files for installation (e.g. launch and bag files, etc.)
-install(PROGRAMS scripts/test_cmd_vel_publisher.py
-  DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION}/scripts
-)
+## Mark other files for installation (e.g. launch and bag files, etc.)
+# install(FILES
+#   # myfile1
+#   # myfile2
+#   DESTINATION ${CATKIN_PACKAGE_SHARE_DESTINATION}
+# )
+
+#############
+## Testing ##
+#############
+
+## Add gtest based cpp test target and link libraries
+# catkin_add_gtest(${PROJECT_NAME}-test test/test_cob_lookat_controller.cpp)
+# if(TARGET ${PROJECT_NAME}-test)
+#   target_link_libraries(${PROJECT_NAME}-test ${PROJECT_NAME})
+# endif()
+
+## Add folders to be run by python nosetests
+# catkin_add_nosetests(test)
