+++ cob_lookat_controller/include/cob_lookat_controller/cob_lookat_driver.h
@@ -61,7 +61,6 @@ public:
 	
 	ros::Subscriber command_vel_sub_;
 	ros::Publisher jointstate_pub_;
-	ros::Publisher debug_jointstate_pub_;
 };
 
 #endif
