+++ cob_lookat_controller/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <package>
   <name>cob_lookat_controller</name>
-  <version>0.5.4</version>
+  <version>0.5.0</version>
   <description>The cob_lookat_controller package</description>
 
   <author email="fxm@ipa.fhg.de">Felix Messmer</author>
@@ -13,6 +13,8 @@
   <build_depend>roscpp</build_depend>
   <build_depend>sensor_msgs</build_depend>
   <build_depend>geometry_msgs</build_depend>
+  <build_depend>std_srvs</build_depend>
+  <build_depend>cob_srvs</build_depend>
   <build_depend>brics_actuator</build_depend>
   <build_depend>kdl_parser</build_depend>
   <build_depend>kdl_conversions</build_depend>
@@ -24,15 +26,20 @@
   <run_depend>roscpp</run_depend>
   <run_depend>sensor_msgs</run_depend>
   <run_depend>geometry_msgs</run_depend>
+  <run_depend>std_srvs</run_depend>
+  <run_depend>cob_srvs</run_depend>
   <run_depend>brics_actuator</run_depend>
   <run_depend>kdl_parser</run_depend>
   <run_depend>kdl_conversions</run_depend>
   <run_depend>orocos_kdl</run_depend>
   <run_depend>tf</run_depend>
   <run_depend>tf_conversions</run_depend>
+  <run_depend>visualization_msgs</run_depend>
+  <run_depend>interactive_markers</run_depend>
 
 
+  <!-- The export tag contains other, unspecified, tags -->
   <export>
-
+    <!-- Other tools can request additional information be placed here -->
   </export>
 </package>
