+++ cob_lookat_controller/scripts/test_cmd_vel_publisher.py
@@ -7,7 +7,7 @@ from brics_actuator.msg import JointVelocities, JointValue
 def start_node():
     rospy.init_node('test_cmd_vel_publisher', anonymous=True)
     #advertise topic
-    pub = rospy.Publisher('/lookat_controller/command_vel', JointVelocities, queue_size=1)
+    pub = rospy.Publisher('/lookat_controller/command_vel', JointVelocities)
     rospy.sleep(1.0)
     
     msg = JointVelocities()
