--- cob_path_broadcaster/package.xml
@@ -1,32 +1,0 @@
-<?xml version="1.0"?>
-<package>
-  <name>cob_path_broadcaster</name>
-  <version>0.5.0</version>
-  <description>
-    This package provides nodes that broadcast tf-frames along various (model-based) Cartesian paths.
-    The tf-frames can be used as targets for the cob_frame_tracker/cob_twist_controller
-  </description>
-
-  <author email="fxm-cm@ipa.fhg.de">Christoph Mark</author>
-  <maintainer email="fxm@ipa.fhg.de">Felix Messmer</maintainer>
-
-  <license>LGPL</license>
-
-
-  <buildtool_depend>catkin</buildtool_depend>
-  
-  <build_depend>roscpp</build_depend>
-  <build_depend>cmake_modules</build_depend>
-  <build_depend>std_msgs</build_depend>
-  <build_depend>tf</build_depend>
-  
-  <run_depend>roscpp</run_depend>
-  <run_depend>rospy</run_depend>
-  <run_depend>std_msgs</run_depend>
-  <run_depend>tf</run_depend>
-
-
-  <export>
-
-  </export>
-</package>
