+++ cob_trajectory_controller/CMakeLists.txt
@@ -2,9 +2,9 @@
 cmake_minimum_required(VERSION 2.8.3)
 project(cob_trajectory_controller)
 
-find_package(catkin REQUIRED COMPONENTS cmake_modules roscpp cob_srvs brics_actuator sensor_msgs std_msgs  actionlib control_msgs dynamic_reconfigure)
+find_package(catkin REQUIRED COMPONENTS roscpp cob_srvs brics_actuator message_generation geometry_msgs sensor_msgs std_msgs  actionlib control_msgs dynamic_reconfigure)
 find_package(orocos_kdl REQUIRED)
-find_package(Eigen REQUIRED)
+find_package(eigen REQUIRED)
 #######################################
 ## Declare ROS messages and services ##
 #######################################
@@ -20,8 +20,8 @@ generate_dynamic_reconfigure_options(
 
 catkin_package(
     DEPENDS orocos_kdl
-    CATKIN_DEPENDS roscpp std_msgs cob_srvs brics_actuator sensor_msgs actionlib control_msgs dynamic_reconfigure
-    INCLUDE_DIRS # TODO include
+    CATKIN_DEPENDS roscpp std_msgs cob_srvs brics_actuator sensor_msgs actionlib control_msgs geometry_msgs message_runtime
+    INCLUDE_DIRS common/include
     LIBRARIES # TODO
 )
 
@@ -29,28 +29,29 @@ catkin_package(
 ## Build ##
 ###########
 
-include_directories(common/include ${catkin_INCLUDE_DIRS} ${orocos_kdl_INCLUDE_DIRS} ${Eigen_INCLUDE_DIRS})
+include_directories(common/include ${catkin_INCLUDE_DIRS} ${orocos_kdl_INCLUDE_DIRS} ${eigen_INCLUDE_DIRS})
+
 
 add_executable(${PROJECT_NAME} ros/src/${PROJECT_NAME}.cpp common/src/genericArmCtrl.cpp common/src/RefValJS_PTP.cpp  common/src/RefValJS_PTP_Trajectory.cpp common/src/TimeStamp.cpp)
-add_dependencies(${PROJECT_NAME} ${catkin_EXPORTED_TARGETS})
+add_dependencies(${PROJECT_NAME} ${PROJECT_NAME}_gencpp)
+#add_dependencies(${PROJECT_NAME} ${PROJECT_NAME}_gencfg)
 
 add_executable(${PROJECT_NAME}_sim ros/src/${PROJECT_NAME}_sim.cpp common/src/genericArmCtrl.cpp common/src/RefValJS_PTP.cpp  common/src/RefValJS_PTP_Trajectory.cpp common/src/TimeStamp.cpp)
-add_dependencies(${PROJECT_NAME}_sim ${${PROJECT_NAME}_EXPORTED_TARGETS})
-add_dependencies(${PROJECT_NAME}_sim ${catkin_EXPORTED_TARGETS})
+add_dependencies(${PROJECT_NAME}_sim ${PROJECT_NAME}_gencpp)
+add_dependencies(${PROJECT_NAME}_sim ${PROJECT_NAME}_gencfg)
 
 add_executable(cob_simulation_tester ros/src/cob_simulation_tester.cpp)
-add_dependencies(cob_simulation_tester ${catkin_EXPORTED_TARGETS})
+add_dependencies(cob_simulation_tester ${PROJECT_NAME}_gencpp)
 
 target_link_libraries(${PROJECT_NAME} ${catkin_LIBRARIES} ${orocos_kdl_LIBRARIES})
 target_link_libraries(${PROJECT_NAME}_sim ${catkin_LIBRARIES} ${orocos_kdl_LIBRARIES})
 target_link_libraries(cob_simulation_tester ${catkin_LIBRARIES} ${orocos_kdl_LIBRARIES})
 
-
 #############
 ## Install ##
 #############
 
-# Mark executables and/or libraries for installation
+## Mark executables and/or libraries for installation
 install(TARGETS ${PROJECT_NAME} 
   ARCHIVE DESTINATION ${CATKIN_PACKAGE_LIB_DESTINATION}
   LIBRARY DESTINATION ${CATKIN_PACKAGE_LIB_DESTINATION}
