+++ cob_trajectory_controller/package.xml
@@ -1,6 +1,6 @@
 <package>
   <name>cob_trajectory_controller</name>
-  <version>0.5.4</version>
+  <version>0.5.3</version>
   <description>This package provides a trajectory controller which controlls velocities for a chain of joints. This controller can be used e.g. with [[schunk_powercube_chain]].</description>
   
   <license>LGPL</license>
@@ -14,13 +14,15 @@
 
   <buildtool_depend>catkin</buildtool_depend>
 
-  <build_depend>cmake_modules</build_depend>
   <build_depend>roscpp</build_depend>
+  <build_depend>eigen</build_depend>
   <build_depend>cob_srvs</build_depend>
   <build_depend>brics_actuator</build_depend>
   <build_depend>sensor_msgs</build_depend>
   <build_depend>actionlib</build_depend>
   <build_depend>control_msgs</build_depend>
+  <build_depend>geometry_msgs</build_depend>
+  <build_depend>message_generation</build_depend>
   <build_depend>std_msgs</build_depend>
   <build_depend>orocos_kdl</build_depend>
   <build_depend>dynamic_reconfigure</build_depend>
@@ -28,10 +30,12 @@
   <run_depend>cob_srvs</run_depend>
   <run_depend>brics_actuator</run_depend>
   <run_depend>sensor_msgs</run_depend>
+  <run_depend>geometry_msgs</run_depend>
   <run_depend>std_msgs</run_depend>
   <run_depend>actionlib</run_depend>
   <run_depend>control_msgs</run_depend>
   <run_depend>orocos_kdl</run_depend>
+  <run_depend>message_runtime</run_depend>
   <run_depend>dynamic_reconfigure</run_depend>
   <!-- <test_depend>roscpp</test_depend> -->
   <!-- <test_depend>cob_srvs</test_depend> -->
