--- cob_twist_controller/CHANGELOG.rst
@@ -1,9 +1,0 @@
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-Changelog for package cob_twist_controller
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-
-0.5.4 (2014-08-26)
-------------------
-* fix dependency-hell on multiple cores
-* moved cob_twist_controller
-* Contributors: Alexander Bubeck, ipa-fxm
