+++ cob_twist_controller/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <package>
   <name>cob_twist_controller</name>
-  <version>0.5.4</version>
+  <version>0.5.0</version>
   <description>The cob_twist_controller package</description>
 
   <author email="fxm@ipa.fhg.de">Felix Messmer</author>
@@ -10,7 +10,6 @@
   <license>LGPL</license>
 
   <buildtool_depend>catkin</buildtool_depend>
-  <build_depend>cmake_modules</build_depend>
   <build_depend>roscpp</build_depend>
   <build_depend>sensor_msgs</build_depend>
   <build_depend>geometry_msgs</build_depend>
@@ -22,8 +21,8 @@
   <build_depend>tf_conversions</build_depend>
 
   
-  <run_depend>roscpp</run_depend>
   <run_depend>rospy</run_depend>
+  <run_depend>roscpp</run_depend>
   <run_depend>sensor_msgs</run_depend>
   <run_depend>geometry_msgs</run_depend>
   <run_depend>brics_actuator</run_depend>
